chore: 단위 변경에 따른 타입 변경

+++ src/main/java/com/tf4/delivery/dto/response/DeliveryRouteLegResponseDto.java
@@ -17,9 +17,9 @@ public class DeliveryRouteLegResponseDto {
     private UUID departureHubId;
     private UUID arrivalHubId;
     private Double estimatedDistanceKm;
-    private Double estimatedTimeMin;
+    private Long estimatedTimeMin;
     private Double actualDistanceKm;
-    private Double actualTimeMin;
+    private Long actualTimeMin;
     private String status;
     private UUID hubDeliveryAgentId;
     private LocalDateTime createdAt;
